refactor: make imports in ddtrace/__init__.py lazy
We refactor the content of the `ddtrace/__init__.py` source to be as
lazy as possible to reduce the side effects for when the library is
simply imported, but not bootstrapped. We move most of the operations
that make more sense only on the bootstrap path to the sitecustomize
instead.
