+++ ddtrace/bootstrap/sitecustomize.py
@@ -13,20 +13,29 @@ Add all monkey-patching that needs to run by default here
 # to allow injecting a custom sitecustomize.py file into the Python process to
 # perform the correct initialisation for the library. All the actual
 # initialisation logic should be placed in preload.py.
-from ddtrace import LOADED_MODULES  # isort:skip
+import ddtrace.bootstrap.cloning as cloning  # noqa
 
-import logging  # noqa:I001
 import os  # noqa:F401
-import sys
-import warnings  # noqa:F401
+import sys  # noqa:F401
+
+# configure ddtrace logger before other modules log
+from ddtrace._logger import configure_ddtrace_logger
+
+configure_ddtrace_logger()  # noqa: E402
+
+# Enable telemetry writer and excepthook as early as possible to ensure we
+# capture any exceptions from initialization
+import ddtrace.internal.telemetry  # noqa: E402
 
 from ddtrace.internal.telemetry import telemetry_writer
 from ddtrace import config  # noqa:F401
 from ddtrace._logger import DD_LOG_FORMAT
 from ddtrace.internal.logger import get_logger  # noqa:F401
-from ddtrace.internal.module import ModuleWatchdog  # noqa:F401
-from ddtrace.internal.module import is_module_installed
-from ddtrace.internal.utils.formats import asbool  # noqa:F401
+
+# TODO(mabdinur): Remove this once we have a better way to start the mini agent
+from ddtrace.internal.serverless.mini_agent import maybe_start_serverless_mini_agent as _start_mini_agent
+
+_start_mini_agent()
 
 # Debug mode from the tracer will do the same here, so only need to do this otherwise.
 if config._logs_injection:
@@ -41,89 +50,10 @@ if config._logs_injection:
 log = get_logger(__name__)
 
 
-if "gevent" in sys.modules or "gevent.monkey" in sys.modules:
-    import gevent.monkey  # noqa:F401
-
-    if gevent.monkey.is_module_patched("threading"):
-        warnings.warn(  # noqa: B028
-            "Loading ddtrace after gevent.monkey.patch_all() is not supported and is "
-            "likely to break the application. Use ddtrace-run to fix this, or "
-            "import ddtrace.auto before calling gevent.monkey.patch_all().",
-            RuntimeWarning,
-        )
-
-
-def cleanup_loaded_modules():
-    def drop(module_name):
-        # type: (str) -> None
-        del sys.modules[module_name]
-
-    MODULES_REQUIRING_CLEANUP = ("gevent",)
-    do_cleanup = os.getenv("DD_UNLOAD_MODULES_FROM_SITECUSTOMIZE", default="auto").lower()
-    if do_cleanup == "auto":
-        do_cleanup = any(is_module_installed(m) for m in MODULES_REQUIRING_CLEANUP)
-
-    if not asbool(do_cleanup):
-        return
-
-    # Unload all the modules that we have imported, except for the ddtrace one.
-    # NB: this means that every `import threading` anywhere in `ddtrace/` code
-    # uses a copy of that module that is distinct from the copy that user code
-    # gets when it does `import threading`. The same applies to every module
-    # not in `KEEP_MODULES`.
-    KEEP_MODULES = frozenset(
-        [
-            "atexit",
-            "copyreg",  # pickling issues for tracebacks with gevent
-            "ddtrace",
-            "concurrent",
-            "typing",
-            "_operator",  # pickling issues with typing module
-            "re",  # referenced by the typing module
-            "sre_constants",  # imported by re at runtime
-            "logging",
-            "attr",
-            "google",
-            "google.protobuf",  # the upb backend in >= 4.21 does not like being unloaded
-            "wrapt",
-        ]
-    )
-    for m in list(_ for _ in sys.modules if _ not in LOADED_MODULES):
-        if any(m == _ or m.startswith(_ + ".") for _ in KEEP_MODULES):
-            continue
-
-        drop(m)
-
-    # TODO: The better strategy is to identify the core modules in LOADED_MODULES
-    # that should not be unloaded, and then unload as much as possible.
-    UNLOAD_MODULES = frozenset(
-        [
-            # imported in Python >= 3.10 and patched by gevent
-            "time",
-            # we cannot unload the whole concurrent hierarchy, but this
-            # submodule makes use of threading so it is critical to unload when
-            # gevent is used.
-            "concurrent.futures",
-        ]
-    )
-    for u in UNLOAD_MODULES:
-        for m in list(sys.modules):
-            if m == u or m.startswith(u + "."):
-                drop(m)
-
-    # Because we are not unloading it, the logging module requires a reference
-    # to the newly imported threading module to allow it to retrieve the correct
-    # thread object information, like the thread name. We register a post-import
-    # hook on the threading module to perform this update.
-    @ModuleWatchdog.after_module_imported("threading")
-    def _(threading):
-        logging.threading = threading
-
-
 try:
     import ddtrace.bootstrap.preload as preload  # Perform the actual initialisation
 
-    cleanup_loaded_modules()
+    cloning.cleanup_loaded_modules()
 
     # Check for and import any sitecustomize that would have normally been used
     # had ddtrace-run not been used.
